+++ centrodefamilia/templates/beneficiarios/beneficiarios_form.html
@@ -185,9 +185,7 @@
             <div class="row mt-4 mb-3">
                 <div class="d-flex justify-content-center">
                     <button id="submit-todos-btn" class="btn btn-primary me-3">Confirmar Preinscripción</button>
-                    <button type="button"
-                            class="btn btn-secondary"
-                            data-page-reload="true">Cancelar</button>
+                    <button type="button" class="btn btn-secondary" data-page-reload="true">Cancelar</button>
                 </div>
             </div>
         </div>
