+++ centrodefamilia/templates/centros/participanteactividad_form.html
@@ -71,9 +71,7 @@
                                             <input type="hidden" name="ciudadano_id" value="{{ c.id }}" />
                                             <button type="submit"
                                                     class="btn btn-sm btn-success"
-                                                    data-submit-with-spinner="Agregando…">
-                                                Agregar
-                                            </button>
+                                                    data-submit-with-spinner="Agregando…">Agregar</button>
                                         </form>
                                     </td>
                                 </tr>
@@ -95,9 +93,7 @@
                         <button id="btn-crear"
                                 type="submit"
                                 class="btn btn-primary"
-                                data-submit-with-spinner="Procesando…">
-                            Crear y Asociar
-                        </button>
+                                data-submit-with-spinner="Procesando…">Crear y Asociar</button>
                     </form>
                 </div>
             </div>
