+++ centrodeinfancia/templates/centrodeinfancia/nomina_form.html
@@ -91,9 +91,7 @@
                         </div>
                         <button type="submit"
                                 class="btn btn-primary"
-                                data-submit-with-spinner="Procesando…">
-                            Crear y Agregar
-                        </button>
+                                data-submit-with-spinner="Procesando…">Crear y Agregar</button>
                     </form>
                 </div>
             </div>
