+++ ciudadanos/templates/ciudadanos/ciudadano_detail.html
@@ -618,15 +618,11 @@
                             <div class="historial-header">
                                 <h6>Historial de Transferencia directa e indirecta</h6>
                                 <div class="periodo-selector">
-                                    <button class="btn"
-                                            data-call-click="cambiarMes"
-                                            data-call-args="[-1]">
+                                    <button class="btn" data-call-click="cambiarMes" data-call-args="[-1]">
                                         <i class="fas fa-chevron-left"></i>
                                     </button>
                                     <span class="mes-actual" id="mes-actual">4 Diciembre 2025</span>
-                                    <button class="btn"
-                                            data-call-click="cambiarMes"
-                                            data-call-args="[1]">
+                                    <button class="btn" data-call-click="cambiarMes" data-call-args="[1]">
                                         <i class="fas fa-chevron-right"></i>
                                     </button>
                                 </div>
