+++ ciudadanos/templates/ciudadanos/ciudadano_form.html
@@ -44,9 +44,7 @@
                 <i class="fas fa-arrow-left mr-1"></i>
                 Volver
             </a>
-            <a href="#"
-               data-history-back="true"
-               class="btn btn-outline-secondary">Cancelar</a>
+            <a href="#" data-history-back="true" class="btn btn-outline-secondary">Cancelar</a>
         </div>
     </div>
     {% if not object %}
@@ -545,8 +543,7 @@
                 </div>
 
                 <div class="section-card" data-section="documentacion">
-                    <div class="section-header"
-                         data-toggle-section-ciudadano="documentacion">
+                    <div class="section-header" data-toggle-section-ciudadano="documentacion">
                         <div class="section-header-left">
                             <div class="section-icon">
                                 <i class="fas fa-camera"></i>
@@ -676,9 +673,7 @@
                         </a>
                     </div>
                     <div>
-                        <button type="button"
-                                data-history-back="true"
-                                class="btn btn-secondary mr-2">
+                        <button type="button" data-history-back="true" class="btn btn-secondary mr-2">
                             <i class="fas fa-times mr-1"></i>
                             Cancelar
                         </button>
