+++ ciudadanos/templates/ciudadanos/grupofamiliar_form.html
@@ -17,9 +17,7 @@
             </a>
         </div>
         <div>
-            <a href="#"
-               data-history-back="true"
-               class="btn btn-secondary">Cancelar</a>
+            <a href="#" data-history-back="true" class="btn btn-secondary">Cancelar</a>
         </div>
     </div>
 
@@ -162,9 +160,7 @@
                         </a>
                     </div>
                     <div>
-                        <button type="button"
-                                data-history-back="true"
-                                class="btn btn-secondary mr-2">
+                        <button type="button" data-history-back="true" class="btn btn-secondary mr-2">
                             <i class="fas fa-times mr-1"></i> Cancelar
                         </button>
                         <button type="submit" class="btn btn-success">
