+++ comedores/templates/comedor/comedor_form.html
@@ -444,9 +444,7 @@
                         {% endif %}
                     </div>
                     <div>
-                        <button type="button"
-                                data-history-back="true"
-                                class="btn btn-secondary mr-2">
+                        <button type="button" data-history-back="true" class="btn btn-secondary mr-2">
                             <i class="fas fa-times mr-1"></i> Cancelar
                         </button>
                         <button class="btn btn-success" type="submit">
