+++ comedores/templates/comedor/nomina_form.html
@@ -110,9 +110,7 @@
                         </div>
                         <button type="submit"
                                 class="btn btn-primary"
-                                data-submit-with-spinner="Procesando…">
-                            Crear y Agregar
-                        </button>
+                                data-submit-with-spinner="Procesando…">Crear y Agregar</button>
                     </form>
                 </div>
             </div>
