+++ core/services/advanced_filters/__init__.py
@@ -1,0 +1,5 @@
+"""Herramientas reutilizables para construir filtros avanzados en listados."""
+
+from .engine import AdvancedFilterEngine
+
+__all__ = ["AdvancedFilterEngine"]
