+++ core/services/column_preferences/__init__.py
@@ -1,0 +1,26 @@
+"""Gestión de preferencias de columnas por listado."""
+
+from .builders import (
+    build_columns_context,
+    build_columns_context_for_custom_cells,
+    build_columns_context_from_fields,
+)
+from .definitions import ColumnDefinition, ColumnResolution
+from .queryset import (
+    apply_queryset_column_hints,
+    build_export_columns,
+    build_export_sort_map,
+)
+from .state import resolve_column_state
+
+__all__ = [
+    "ColumnDefinition",
+    "ColumnResolution",
+    "apply_queryset_column_hints",
+    "build_columns_context",
+    "build_columns_context_for_custom_cells",
+    "build_columns_context_from_fields",
+    "build_export_columns",
+    "build_export_sort_map",
+    "resolve_column_state",
+]
