+++ core/services/favorite_filters/__init__.py
@@ -1,0 +1,20 @@
+"""Utilidades para filtros avanzados favoritos."""
+
+from .config import (
+    ConfiguracionFiltrosSeccion,
+    SeccionesFiltrosFavoritos,
+    TTL_CACHE_FILTROS_FAVORITOS,
+    clave_cache_filtros_favoritos,
+    obtener_configuracion_seccion,
+)
+from .validation import normalizar_carga, obtener_items_obsoletos
+
+__all__ = [
+    "ConfiguracionFiltrosSeccion",
+    "SeccionesFiltrosFavoritos",
+    "TTL_CACHE_FILTROS_FAVORITOS",
+    "clave_cache_filtros_favoritos",
+    "normalizar_carga",
+    "obtener_configuracion_seccion",
+    "obtener_items_obsoletos",
+]
