+++ core/trash_views.py
@@ -141,6 +141,73 @@ def _instance_details(instance):
     return details
 
 
+def _build_filter_form_and_data(request, model_choices, models_by_key):
+    form_data = request.GET.copy() if request.GET else None
+    if form_data is not None:
+        selected_key = form_data.get("model", "")
+        if selected_key not in models_by_key:
+            selected_key = ""
+        form_data["model"] = selected_key
+        filter_form = TrashFilterForm(form_data, model_choices=model_choices)
+    else:
+        selected_key = ""
+        filter_form = TrashFilterForm(
+            model_choices=model_choices,
+            initial={"model": selected_key},
+        )
+
+    form_is_valid = filter_form.is_valid() if filter_form.is_bound else True
+    filter_data = (
+        filter_form.cleaned_data if filter_form.is_bound and form_is_valid else {}
+    )
+    return filter_form, selected_key, filter_data
+
+
+def _get_models_to_scan(selected_key, models_by_key, order_keys):
+    if selected_key:
+        return [models_by_key[selected_key]]
+    return [models_by_key[key] for key in order_keys]
+
+
+def _build_deleted_rows(models_to_scan, filter_data):
+    rows = []
+    for model in models_to_scan:
+        queryset = model.all_objects.filter(deleted_at__isnull=False).select_related(
+            "deleted_by"
+        )
+        if filter_data:
+            queryset = _search_queryset(model, queryset, filter_data.get("q", ""))
+            queryset = _apply_deleted_by_filter(queryset, filter_data.get("deleted_by"))
+            queryset = _apply_deleted_date_range_filter(
+                queryset,
+                filter_data.get("deleted_from"),
+                filter_data.get("deleted_to"),
+            )
+
+        model_label = str(model._meta.verbose_name).title()
+        for item in queryset.order_by("-deleted_at", "-pk"):
+            rows.append(
+                {
+                    "obj": item,
+                    "app_label": item._meta.app_label,
+                    "model_name": item._meta.model_name,
+                    "model_label": model_label,
+                }
+            )
+
+    rows.sort(
+        key=lambda row: (row["obj"].deleted_at, row["obj"].pk),
+        reverse=True,
+    )
+    return rows
+
+
+def _build_pagination_querystring(request):
+    pagination_query_params = request.GET.copy()
+    pagination_query_params.pop("page", None)
+    return pagination_query_params.urlencode()
+
+
 class TrashListView(LoginRequiredMixin, SuperAdminRequiredMixin, View):
     """List soft-deleted records by model with basic search."""
 
@@ -154,22 +221,10 @@ class TrashListView(LoginRequiredMixin, SuperAdminRequiredMixin, View):
 
         models_by_key = _soft_models_by_key()
         order_keys = [key for key, _ in model_choices]
-        form_data = request.GET.copy() if request.GET else None
-        if form_data is not None:
-            selected_key = form_data.get("model", "")
-            if selected_key not in models_by_key:
-                selected_key = ""
-            form_data["model"] = selected_key
-            filter_form = TrashFilterForm(form_data, model_choices=model_choices)
-        else:
-            selected_key = ""
-            filter_form = TrashFilterForm(
-                model_choices=model_choices, initial={"model": selected_key}
-            )
-
-        form_is_valid = filter_form.is_valid() if filter_form.is_bound else True
-        filter_data = (
-            filter_form.cleaned_data if filter_form.is_bound and form_is_valid else {}
+        filter_form, selected_key, filter_data = _build_filter_form_and_data(
+            request=request,
+            model_choices=model_choices,
+            models_by_key=models_by_key,
         )
         selected_key = filter_data.get("model", selected_key or "")
         selected_model_label = (
@@ -178,55 +233,27 @@ class TrashListView(LoginRequiredMixin, SuperAdminRequiredMixin, View):
             else "Todos los modelos"
         )
 
-        models_to_scan = (
-            [models_by_key[selected_key]]
-            if selected_key
-            else [models_by_key[key] for key in order_keys]
+        rows = _build_deleted_rows(
+            models_to_scan=_get_models_to_scan(
+                selected_key=selected_key,
+                models_by_key=models_by_key,
+                order_keys=order_keys,
+            ),
+            filter_data=filter_data,
+        )
+        page_obj = Paginator(rows, self.paginate_by).get_page(
+            request.GET.get("page") or 1
         )
 
-        rows = []
-        for model in models_to_scan:
-            queryset = model.all_objects.filter(
-                deleted_at__isnull=False
-            ).select_related("deleted_by")
-            if filter_data:
-                queryset = _search_queryset(model, queryset, filter_data.get("q", ""))
-                queryset = _apply_deleted_by_filter(
-                    queryset, filter_data.get("deleted_by")
-                )
-                queryset = _apply_deleted_date_range_filter(
-                    queryset,
-                    filter_data.get("deleted_from"),
-                    filter_data.get("deleted_to"),
-                )
-            for item in queryset.order_by("-deleted_at", "-pk"):
-                rows.append(
-                    {
-                        "obj": item,
-                        "app_label": item._meta.app_label,
-                        "model_name": item._meta.model_name,
-                        "model_label": str(model._meta.verbose_name).title(),
-                    }
-                )
-
-        rows.sort(
-            key=lambda row: (row["obj"].deleted_at, row["obj"].pk),
-            reverse=True,
-        )
-
-        paginator = Paginator(rows, self.paginate_by)
-        page_obj = paginator.get_page(request.GET.get("page") or 1)
-        pagination_query_params = request.GET.copy()
-        pagination_query_params.pop("page", None)
-        pagination_querystring = pagination_query_params.urlencode()
-
-        active_filters = []
-        if filter_data:
-            active_filters = _build_active_filters(
+        active_filters = (
+            _build_active_filters(
                 filter_data,
                 selected_model_label=selected_model_label,
                 default_model_key="",
             )
+            if filter_data
+            else []
+        )
 
         context = {
             "filter_form": filter_form,
@@ -237,7 +264,7 @@ class TrashListView(LoginRequiredMixin, SuperAdminRequiredMixin, View):
             "rows": page_obj.object_list,
             "search_query": filter_data.get("q", ""),
             "active_filters": active_filters,
-            "pagination_querystring": pagination_querystring,
+            "pagination_querystring": _build_pagination_querystring(request),
             "current_full_path": request.get_full_path(),
             "show_all_models": not selected_key,
         }
