+++ templates/includes/new_base.html
@@ -130,8 +130,7 @@
             </div>
             {% include "scripts/mainscripts.html" %}
             {% include "includes/scripts/toastr_messages.html" %}
-            {% block customJS %}
-            {% endblock %}
+            {% block customJS %}{% endblock %}
         </div>
     </body>
 </html>
